Unify komenci and classic verification flows (#5)
### Disclaimer

First of all, yeah, it's big, but it's NOT that big really 🙂 Cause, for example, `Settings.test.tsx.snap` alone brings +2105 LOC from the total 4k. I've tried, but was not able to split this PR into multiple ones, so please bear with me here!

### Description

The initial implementation of feeless verification (using Komenci service) was done as a copy-paste from classic flow (when user pays for verification from their own funds) with adjusting logic, when needed. This has introduced significant amount of duplicated code, hence technical debt. This PR aims to remove it.

It's also a first part of the bigger effort to refactor verification in general.

#### Approach

First, I've removed everything related to `fetchVerificationState` in both feeless and classic flows. So, now we do not fetch the whole verification state every time something has changed, but rather maintain it locally with incremental changes. Then I've implemented the proposed state machine (see diagram below) with this removing feeless verification code, which was duplicated. 

Also, I've start to migrate verification logic from `src/identity` to `src/verify` as it grown really big and I think it makes sense to separate it.

I'm up to schedule a session, where I can walk people through the PR, if that would be beneficial. Please react to this message with 👍🏻 if you are interested in this.

#### State machine
Here is a diagram of the state machine, which is implemented in the PR. Pay attention, that everything below `FetchingOnChainData` is not implemented yet, but will be address in the subsequent PRs.

I've used http://www.plantuml.com/ diagram to build it for a couple of reasons:
* It's free
* Images are build form code, so we can efficiently maintain relevancy of this diagram
* It's way easier than use https://xstate.js.org/viz/, which I used initially. Also, cause we do not use the library, thereis no real benefit to used their syntax.

![State machine diagram](https://raw.githubusercontent.com/celo-org/wallet/i1skn/refactor-feeless-verification/packages/mobile/docs/verification_state_machine.png)

### Other changes

I've tried https://redux-toolkit.js.org/ as a part of this PR and must say it's way faster to iterate on code compare to our current approach. For example, this is how actions are created:
```
// redux-toolkit
const startVerification = createAction<{ e164Number: string; withoutRevealing: boolean }>('IDENTITY/START_VERIFICATION')

// --------------------------------

// current way
enum Actions {
  START_VERIFICATION = 'IDENTITY/START_VERIFICATION',
  ...
}

interface StartVerificationAction {
  type: Actions.START_VERIFICATION
  withoutRevealing: boolean | undefined
  e164Number: string
}

export type ActionTypes =
  | StartVerificationAction
  ...

const startVerification = (
  e164Number: string,
  withoutRevealing: boolean | undefined
): StartVerificationAction => ({
  type: Actions.START_VERIFICATION,
  withoutRevealing,
  e164Number,
})

```

### Tested

Yes

### Related issues

- Fixes https://github.com/celo-org/celo-monorepo/issues/6094

### Backwards compatibility

Yes