Fix the fetching of exchange rates (#255)
### Description

We had a bug in the fetching of the exchange rates that made us fetch the whole thing every time there was an update, which is every 30 minutes. The number of elements to fetch each time depended on how much time it passed since the last use of the app, for new users and for users that didn't open the app in a couple months we would load all the history for three months.

We were also fetching the whole exchange rate history from Firebase and filtering it in memory instead of only fetching the latest three months. We tried to filter, but we were sending `1` as `startAt` which returns everything.

With the current changes, we load everything since the max between the last update and three months ago and then we listen only for new items, we don't reload everything again.

### Other changes

- Added a migration to delete the exchange rate history from Redux since it was likely that it was repeated a bunch of times.

### Tested

Mostly manually, by making changes and adding and modifying items in the RTDB to see that the behavior was the expected one.

### Related issues

- Fixes https://github.com/celo-org/celo-monorepo/issues/6284

### Backwards compatibility

N/A