Notification Service: Only update last notified block if an event was found (#244)
### Description

There was an error on the notification service where we were automatically adding one to the last block notified if we didn't find anything. This automatic addition was causing us to not send notifications because the last notified block was much higher than the last block.

### Other changes

Added a test for the invite poller.

### Tested

Manually and added a test. This has already been deployed.

### Related issues

- Fixes #

### Backwards compatibility

N/A