Show contact names on pending Escrow notification and list (#288)
### Description

Note: Out of the 1000+ new lines 800 are a snapshot file.

There are bigger changes to the escrow invite flow coming but the focus of this PR is to just show the correct name of the invited contacts in the Pending Escrow list and in the notification that shows up in the `NotificationBox`.

To do this, I added a new mapping to de `identity` reducer called `identifierToE164Number`. The identifier can be easily calculated with the phone number and the pepper and we already have a `e164NumberToSalt` (we should rename salt -> pepper btw) mapping so we can populate this new mapping from the existing one and new items will be added to both mappings going forward.

The reason we need this is that escrow invites are sent using the identifier so we need to get a contact from this identifier. This is not working right now in production (it shows Unknown for invited people) and previously it worked by doing the mapping using the temporary address we used to create for invites but we don't do that anymore since Komenci was released so it hasn't worked since then.

### Other changes

- Fired the `refreshAllBalances` action when `WalletHome` is mounted so that the escrow payments are refreshed. Otherwise you had to pull to refresh for the info to get refreshed.
- Escrow invites now show up immediately after sending them and can be reclaimed.

### Tested

Manually on a physical device. Also added some unit tests.

### Related issues

- Fixes #287

### Backwards compatibility

N/A