docs: improve templates and create template config with external links (#281)
### Description

This is an update to our issue templates including a more detailed bug report template, a feature request template, and a `config.yaml` to house external links when users go to submit an issue. This is to help bin things more appropriately and prevent support requests from being submitted as tickets. Also it includes a link to how to submit a security related issue and many other CELO & cLabs links in the `config.yaml`

### Tested

The changes were tested by viewing the `.md` files in preview mode. This will not need to be tested in the release cycle as it doesn't impact application code; however, product review is needed be help to confirm stylistic wording.

### Backwards compatibility

These changes should be backwards compatible and only impact new issues.