Display CIP8 profile info in Transaction Feed (#37)
 migrated from https://github.com/celo-org/celo-monorepo/pull/6605

### Description
This PR includes a refactor of how we handle recipients. When handling recipient info, the paradigm is to put all information into a Recipient object which is passed between components/functions. This parameter is made mandatory, instead of individually passing address/number/name/thumbnail variables individually.

When a user sends a transaction, it uploads symmetric keys which grants the recipient permission to access their profile information. When the recipient receives the transaction, it downloads and decrypts this profile info, saving it in redux.

Transactions in the feed displays CIP8 profile information if it is available for the recipient. This PR does not handle contact list and send flow UI.

### Other/Drive-by Changes
Removed recipients from the persist blacklist, because I want to persist valoraRecipientsCache. Previously the only variable in the recipients reducer was the recipientCache (renamed to phoneRecipientCache ), which I guess was blacklisted because we reconstruct it at the start of each app session. Not sure what the impact of this is

### Tested
Sending transactions between Android device and iOS emulator

### Related issues
Fixes #18 partially