Initialize sentry earlier (#326)
### Description

Moved sentry init earlier in order to catch errors that might happen during redux store loading.

### Other changes

Also updated the getStoredState function to not return empty strings.

The actual error that was happening in production was the database throwing an error like:

> 05-04 18:11:12.912 24367 24456 I ReactNativeJS: redux/store :: Failed to retrieve redux state. :: Couldn't read row 0, col 0 from CursorWindow.  Make sure the Cursor is initialized correctly before accessing data from it. in Error: Couldn't read row 0, col 0 from CursorWindow.  Make sure the Cursor is initialized correctly

This happend because when the store reached a size of about 2MB the read would fail. The way the redux library stores the data is all in one row in the database, which is not great for our use case. Anyway we have moved away from using AsyncStorage since then but just in case any error, such as during a migration happens in the future we will hopefully catch it early before shipping to production.

### Tested

Manually. Threw an exception in store.getStoredState and noticed it does not get captured by sentry because it is not initialized at the time.


### Related issues

- Fixes #327 

### Backwards compatibility

Yes.