Turn Send screen into a functional component (#411)
### Description

This PR is part 1 of https://github.com/celo-org/wallet/pull/406

This PR refactors the Send screen to be a functional component rather than a class component to make it easier to review the actual changes in the other PR. Unfortunately, the nature of the changes make reviewing the Send screen specifically pretty difficult. The only real things that changed here are:
- Added SendHeader to replace the react-navigation provided header (more on this below).
- Re-added the verification banner on the Send screen that we removed during the CB earn campaign (and deleted the TODOs). Not hiding the contacts though for now.

#### The Send screen header
The Send screen now has a custom header. Created a `CustomHeader` component for this. The reason is that using the header provided by react-navigation like we've been doing causes a bunch of navigation issues. The screen has an animation where it shows up from the bottom. To achieve that we needed to put the Send screen in a weird place in the navigation tree because otherwise the header comes up from the top instead. Since the screen wasn't in the stack navigation object, navigating from the Send screen to another screen caused the send screen to disappear from the stack, so if the user navigated to somewhere else (SendAmount or on a CTA in the list header) and back the user actually returned to the Home screen.
Having this custom header fixes these issues without negative consequences that I could find.

### Other changes

N/A

### Tested

Manually mostly.

### How others should test

- Open the app, tap the `Send` button.
- Check that the list of contacts works well, it filters the contacts correctly when typing in the search input.
- Go to the next screen (`SendAmount`) by picking a contact and back. See that you're back on the Send screen.

### Related issues

- Part of #405

### Backwards compatibility

N/A