Don't start verification saga unless user starts verification (#454)
### Description

The verification saga was being started on app open, so it was expecting the user to get verified within 10 minutes. If that didn't happen a 'Confirmation Timed Out' error message showed up. Changed so that the timeout is only waited on when the `start` button is pressed in the verification screen.

### Other changes
Added in the `verification_start` analytics event that was removed on accident.

### Tested

Manually looking at the logs and waiting.

### How others should test

Open the app, do other stuff but make sure the app is open at least 10 minutes. See that there is no error message that says 'Confirmation Timed Out'.

### Related issues

- Fixes #

### Backwards compatibility

N/A