Move payment request notifications to Cloud Function (#430)
### Description

Moved the payment requests notification sending from the notification service to a database-triggered Cloud Function.

See https://www.notion.so/clabsco/Notification-Service-Indexer-064ca4f58be64c3384ac8fb5f00fca56 for more information on the motivation for doing this.

### Other changes

N/A

### Tested

Manually and added unit test. This is deployed in Alfajores.

### How others should test

Send a payment request to an address. See that the requested address receives a push notification.

### Related issues

- Fixes #428

### Backwards compatibility

N/A