Show warning about sending a test transaction when sending to new addresses (#464)
### Description

When a user selects a raw address to send to, a warning about sending a test transaction is show before they proceed. The user has the option to turn off this warning forever. Currently there is no way to re-enable the warning after it has been disabled

### Other changes

A recipient is added to the recent recipients list when a payment/invite is sent to them, not just when they have been selected from the RecipientPicker

### Tested

tested in emulator

![Screenshot_1623198507](https://user-images.githubusercontent.com/15326838/121274996-ae9cdf80-c899-11eb-9866-1ac53283243c.png)

### How others should test

When selecting a recipient, type in an address and select it. The warning does not appear for recipients selected from the recent recipient list

### Related issues

- Fixes #250 

### Backwards compatibility

yes