Tomm/e2e ramps (#451)
### Description

Increases the automated coverage of our on and off ramps in the Valora. Tests focus on in app behavior along with checking integration with 3rd party providers. PNG snapshots are used to confirm that 3rd navigating away from the app behaves as expected.

### Other changes

- [x] Updates Jest Snapshots with new testID's.
- [x] Add associated png snapshots for Android and iOS.

### Tested

- [x] All tests running locally on iOS and Android devices matching Circle Ci Configurations
- [x] Run in Circle Ci 
- [x] Individual specs run locally

### How others should test

* Navigate to the `wallet/packages/mobiles` directory.

* All Tests can be run

  ```sh
  # Using Multiple Emulated Devices
  yarn test:e2e:ios -w 4
  
  # Runs One Emulated Device
  yarn test:android -w  1
  ```

* A Specific spec.js can be run 

  ```sh
  yarn test:e2e:ios -w 1 -f OnRamps.spec.js
  ```

* Rerunning with out rebuilding

  ```sh
  # Both 'OnRamps.spec.js' and 'OffRamps.spec.js' will run
  yarn test:e2e:ios -w 2 -d -f Ramps.spec.js
  ```

This does not need to be tested by QA, but monitoring build times prior to merge to main will be important as the asset folder grows.

### Backwards compatibility

This PR can be merged alone and also contains all of the contents of PR #260.