Properly handling WalletConnect deep/universal links (#459)
### Description

Take a look [here](https://docs.walletconnect.org/v/2.0/mobile-linking) for an overview of the problem. Basically we're conforming to the spec for deep/universal links so it's easier for other frontend libraries to integrate with Valora.
 
### Other changes

N/A

### Tested

Added unit tests.

### How others should test

Can use Plock.fi and use-contractkit.vercel.app for testing. These changes should make it smoother to use on mobile.

### Related issues

N/A