Fix crash due to redux immutability (#838)
### Description

This change is fixing this crash:
https://sentry.io/organizations/celo/issues/2499735012/?project=1250733&query=is%3Aunresolved

The crash was introduced on this PR: https://github.com/celo-org/wallet/pull/656 when switching to redux toolkit. The new toolkit uses `immer` under the hood which makes the objects immutable and  we were using Object.assign in some places 

### Other changes

N/A

### Tested

Manually

### How others should test

Use the app to send to some people so that the recipients cache is populated and then just try to send to someone. The app should never crash.

### Related issues

- Fixes #837

### Backwards compatibility

N/A