Fix clock sync (#868)
### Description

Looking at the logs provided in https://github.com/celo-org/wallet/issues/544 I noticed that when the clock sync check fails the drift is always 20 seconds. Then looking at the ntp library we are using here https://github.com/artem-russkikh/react-native-ntp-client/blob/master/lib/ntp-client.js it looks like the default timeout for the request is 10 seconds. Because we attempt 2 requests to 2 different ntp servers these add up to 20 seconds. Then the check fails because we want to be within 4 seconds. So if something is wrong with the network such as packet loss or udp port being blocked (which is pretty common on mobile devices) the check would always fail. Now it would be interesting to see what actually caused the timeout in this case.

The fix sets the ntp timeout to 2 seconds.

Of course there might be other issues causing this, such as the one we discussed where we don't account for the network delay roundtrip in the calculation.

Update: Looking at the history of this util function we used to use react-native-clock-sync which uses underneath react-native-ntp-client. The clock-sync lib keeps track of historical requests in order to estimate and adjust for network round trip delay. Using the ntp client directly does not do that as it is just a util that talks to ntp directly without any state.

### Other changes

I also modified the getRemoteTime function to return the initial start time in case of a timeout.

### Tested

In order to block the time servers we specify I modified the hosts file on my android emulator using the following steps roughly: block the time servers, then restart the app and notice after 10 seconds the clock sync screen shows up. There are probably other ways to test this but this is what I did locally using an android emulator:

Create a hosts file that looks something like:

`127.0.0.1       localhost`
`::1             ip6-localhost`
`172.0.0.1	time.google.com`
`172.0.0.1	time.cloudflare.com`
`

Now in order to push this file to the device we need to root the device.

> adb root
> adb disable-verity
> adb reboot
> adb root
> adb remount
> adb push hosts /etc/hosts

Now restart the app and wait 10 seconds to see the SetClock screen.

### How others should test

Start the app in various network conditions from really good wifi to bad signal mobile. Make sure the set clock screen does not show up after 5 seconds.

### Related issues

- Fixes #544 

### Backwards compatibility

Yes.