Make unencrypted comment warning text more visible (#639)
### Description

@gtank and I noticed that its way too easy to miss the the warning that a comment is being senty
publically in the clear. As a remedy, this PR makes the fairly simple change if making the warning
text `colors.warning` (red) and slightly larger.

<img alt="Screenshot_before" src="https://user-images.githubusercontent.com/5684286/123715086-70f1fd80-d82c-11eb-8b43-5a8b1f935f68.png" width="40%"/> <img alt="Screenshot_after" src="https://user-images.githubusercontent.com/5684286/124331517-32b65000-db44-11eb-8387-260189c8bf70.png" width="40%"/> 

### Tested

Navigated to the send flow with a valid addrss that is not registered on Alfajores and entered a
comment. Observed that the warning text had changed.

### How others should test

Navigate to the send flow with a valid addrss that is not registered on Alfajores. Check that the
warning text has chnaged.

### Backwards compatibility

No backwards compatibility concerns

### Related issues

* https://github.com/celo-org/wallet/issues/61
* https://github.com/celo-org/wallet/issues/205