Update debug logging redux blocklist (#891)
### Description

In the debug logger, used in development mode and not in production, the app will log all actions
sent to the Redux saga processor. In order to avoid logging PII (and to reduce the noise in the
logs) there is a blocklist of events which should not have their payloads logged. This list was
created referencing the event types as strings, and so does not raise a compiler error if one of
those types is changed or removed. As a result, the blocklist has gotten out of date, and some
information such as the user's contact list.

### Other changes

* Changed "blacklist" to "blocklist" where possible in the source code.
* Removed the REDEEM_INIVTE event from the blocklist, as it no longer exists. Removed in 145534020927251e3b14198a41347d979588330d

### Tested

Manually tested these changes by observing the payloads were not logged, as intended.