Report availability of Google and Huawei Mobile Services (#892)
### Description

We plan to use Google Drive on Android as a component in encrypted cloud backup for account keys,
but in order to use Drive, the user needs to have a Google account and Google Mobile Services (i.e. Google Play Services)
installed on their device. Some Android devices, especially those produced by Huawei since 2019, are
shipped without Google Mobile Services, and the user on these devices is unlikely to have a Google
account.

In order to determine how many Valora users are on Android devices without Google Mobile Services,
this PR introduced functionality to check whether Google Mobile Services are available on startup.
This information is recorded in the state tree and sent in an event to Segment. Sending the event
to Segment will allow us to determine what portion of users will be able to utilize any new feature,
including cloud backup, that will rely on Google Mobile Services.

Additionally, this PR adds a check for Huawei Mobile Services, which is an alternative to Google Mobile Services provided by Huawei on their devices (since they cannot ship with Google integration either in mainland China or, since 2019, abroad)

### Other changes

* Removed unused permission function `requestPhoneStatePermission `

### Tested

* Ran the app and observed the logs for the expected events.

### Related issues

- Related: https://github.com/celo-org/wallet/issues/846