Add push notifications for invite rewards (#1014)
### Description

Send a push notification to users when an invite is redeemed and they get the reward for inviting someone.

### Other changes

Removed the previous invite poller that simply send a notification when anyone you invited redeemed their invite. Should be replaced by this one in almost all cases.

### Tested

In Alfajores

### How others should test

Hard to test, but invite a new phone number, verify it and check that the inviter receives a push notification when they receive the reward. 

### Related issues

- Fixes #1013

### Backwards compatibility

N/A