Return EOA address on WalletConnect (#1074)
### Description

In the WalletConnect integration we are returning the MTW address instead of the EOA. This is super dangerous! Users that connect using WalletConnect might end up making transactions to the MTW address and recovering the funds might be a bit painful (they wouldn't be lost, but it would be annoying).

### Other changes

N/A

### Tested

Using [use-contractkit](http://use-contractkit.vercel.app/)

### How others should test

Hard to test because [the public version of use-contractkit](http://use-contractkit.vercel.app/) doesn't work (I tested with a local version with some minor changes). If it did work one could simply connect your account in that website and check that the address listed is the EOA.

### Related issues

- Fixes #1075

### Backwards compatibility

N/A