Add Cloud Function to fetch account <-> wallet address mapping (#1077)
### Description

Added a Cloud Function that allows us to fetch the account address(es) that are linked to a wallet address. This will be used in a tool that will allow users to transfer funds from their MTW to their EOA if they sent funds to the MTW my mistake. This happened to some users recently because of a bug in the WalletConnect integration.

### Other changes

N/A

### Tested

Manually using the tool mentioend above


### How others should test

If you need access to the tool let me know, but this shouldn't be a part of our QA release testing

### Related issues

- See https://valora-app.slack.com/archives/C01D6H6116V/p1630542810004200

### Backwards compatibility

N/A