Fix Google Play Services check in release builds (#1109)
### Description

The check for GMS fails in release builds but works in dev builds. This usually points to something related to proguard. A method that was needed by this library was getting stripped by proguard in release mode, so adding that to the rules fixes it.

Follow issue from the library here: https://github.com/react-native-device-info/react-native-device-info/pull/1302

### Other changes

Also upgraded the library itself thinking that might fix it, but it didn't. On the issue they are discussing a different fix so worth keeping an eye on it.

### Tested

Enabled minify in debug mode (in build.gradle) and looked for the "Google Mobile Services" log.

### How others should test

Run an emulator with google play services installed and then run:
`adb logcat | grep "Google Mobile Services" ` 

### Related issues

- Fixes #1046 

### Backwards compatibility

Yes.