Adding editorConfig and workspace files (#1181)
### Description
I've noticed that we all develop in different editors and IDEs, often times I make unintentional style changes when I touch a file. I'm adding a vscode-workspace config file and .editorConfig file in this PR.  

EditorConfig https://editorconfig.org/ - It helps to maintain consistent style for developers across different editors and IDEs. It is supported by most common editors and it enables editors to read the file format and adhere to the code styles.

vscode-workspace https://code.visualstudio.com/docs/editor/workspaces - This is a vscode specific config that offers a bunch of vscode specific settings for the project. It also would recommend users to install a few extensions such as eslint when they open the workspace for the first time. I found this feature to be potentially helpful for anyone who tries to setup a new Valora env in vscode.

### Tested

n/a