Reading known addresses in blockchain-api (#1130)
### Description

Avoid reading known addresses from firebase in the client, but receiving that info from the blockchain-api

### Tested

Changes were tested running the app locally. But also with unit and e2e tests.

### Related issues

* Fixes #426

### Backwards compatibility

With these changes, we are not showing  anymore known addresses while receiving notifications, since it's not coming from the blockchain-api, that will be done in another PR. Probably the best would be to wait until the other PR is ready to merge to merge both at the same time so there is no unexpected behavior.

Note: If this PR seems big, I have created another [PR](https://github.com/valora-inc/wallet/pull/1110) with most of the changes on the mobile side (but it might be a bit outdated after some changes I did in this PR). However, both PRs should be merged together.