Reenable WalletConnect v2 (#1223)
### Description

This is a hotfix that was done on the release branch to re-enable WC v2. opening a PR to merge to main as well.

### Other changes

N/A

### Tested

Manually, there's been a lot of discussion on Slack, especially on the #clabs-eng-walletconnect channel and this thread: https://valora-app.slack.com/archives/C014SRL4URY/p1631211271054000

### How others should test

Test the WalletConnect v2 flow and see that it works. For now it can be tested from https://use-contractkit-git-alexbharley-fix-wc-2-alexbharley.vercel.app/ 

### Related issues

- Fixes #

### Backwards compatibility

N/A