Clean up in-app notification showing logic (#1231)
### Description

Cleaned up the logic to decide when the `NotificationBox` should show up or not. The logic was already implemented in the `NotificationBox` itself but was repeated with a bug on the `WalletHome`. I just removed the check from `WalletHome`.

### Other changes

Removed some old code around section headers (none of them have headers)

### Tested

Manually locally.

### How others should test

Dismiss all non-remote notifications and check that remote notifications still show up.
You should also go through the CELO education screens since that was also required before for remote notifications not to show up 🍝  

### Related issues

- Fixes #1230

### Backwards compatibility

N/A