Spruce/log phone type analytics (#1271)
## Description
This logs the type of phone number to the logging service, such as voip or mobile to determine high level statistics around the account demographics. This is useful to identify if the network should be concerned about sybil attacks by an increase in cheap-to-generate voip numbers. These changes are related and companion to celo-org/celo-monorepo#8286, which adds the verifiable credential to the attestation service response.

Note the original PR for this change can be found here https://github.com/valora-inc/wallet/pull/879

## Other changes
Feature flag to enable/disable.

## Tested
We add tests for the data schema.
These changes were difficult to test E2E without setting up a Firebase instance for Valora and enabling analytics.
How others should test
Does this need to be tested by QA in the next release cycle? If so please give a brief explanation of how to test these changes.

## Related issues
Companion to Allow Attestation Service to Issue W3C Verifiable Credentials for Phone Number Type celo-org/celo-monorepo#8286
Addresses Verification observability #685

## Backwards compatibility
We have added a feature flag to turn this on or off, allowing for graceful degradation.