Add multi-token support to the SendAmount screen (#1302)
### Description

Third PR of the send flow changes for multi-token support.
Depends on https://github.com/valora-inc/wallet/pull/1303

I know it looks big, but a lot of the LOC are snapshot updates or tests. It's hard to separate this further but if someone has strong thoughts on it I'm open to hearing them.

Recommendation for order of review:
- Start in `src/send/SendAmount/index.ts`, that's the main file of the SendAmount screen. From there, there are different parts in different files:
- `SendAmountHeader` has the header, from where users can change the token they want to send.
- `TokenPickerSelector` is a small UI component that is used in `SendAmountHeader`
- `SendAmountValue` has the input section, with the number, the max and the swap buttons that allows users to input in fiat or crypto.
- `useTransactionCallbacks` has the `onSend` and `onRequest` callbacks.

Work remaining for #1126:
- Changes to SendConfirmation screen and send saga to send in any erc20 token.
- Take into account the gas fee in some cases when pressing the MAX button.
- Make sure the payment request, invite and paying through QR code flows work correctly.
- Show the pending transfer in the feed. Will likely require working with #1262 implementation.


https://user-images.githubusercontent.com/6062888/138405636-0bcfd8b6-25d3-470a-b8c3-3974f82b7270.mov


### Other changes

- N/A

### Tested

Manually and with unit tests

### How others should test

Not yet, the whole flow needs to be tested after it's done. For now just going to the send amount screen and playing around with all the buttons in the UI.

To get to the screen you need to make one change to the `Send` screen. In line 128, the `navigate` needs to go to `Screens.SendAmount` and you need to comment out the `forceCurrency` param. After that the new screen can be opened by going through the Send flow normally 

### Related issues

- Part of #1126

### Backwards compatibility

N/A