Send confirmation screen (#1312)
### Description

Added multi-token support to the `SendConfirmation` screen. The last step missing in order to be able to test e2e is adding multi-token support to the send saga and the simple send use-case will be able to be tested e2e.

Of the total LOC changes, 1110 are a single snapshot file. The PR really is not as big as it seems :(

The main file in this PR is `SendConfirmation`. That's a good place to start the review.
I also copied `TotalLineItem` into `TokenTotalLineItem` to have the token-based version of that file. 

https://user-images.githubusercontent.com/6062888/138785634-405943a9-0003-4330-a7c7-6f0634114ada.mov

### Other changes

N/A

### Tested

Manually and with unit tests.

### How others should test

Wait until the next PR and you'll be able to test the whole send with tokens flow :)

### Related issues

- Part of #1126

### Backwards compatibility

N/A