Specify nonce for exchange transactions (#1268)
### Description

There seemed to be a race condition happening, where the node that executed exchange transactions was not aware of the approve transaction that was executed prior, which was causing the exchange transaction to fail. By fetching the account nonce and specifying the nonce in the exchange transaction, this should prevent this situation.

more context here: https://celo-org.slack.com/archives/CP5V8KF51/p1631808226092700

### Other changes

`sendAndMonitorTransaction` not accepts nonce as a parameter

### Tested

exchanged cUSD for CELO on alfajores


### How others should test

Try to execute an exchange transaction (ie. buying or selling CELO with a stable asset)

### Related issues

- Fixes #1043

### Backwards compatibility
yes