Change WalletHome to be a function component (#1423)
### Description

Changed the `WalletHome` component to be a function component. This is the main screen of the app and was due for the update. No logic should have been changed.

### Other changes

- Changed in some places the wording of `gold` for `celo`. Gold is the original name of the token but was changed a longtime ago. Didn't go through all the code to change it, only the places that were affected by this PR.

### Tested

Manually and updated the unit tests.

### How others should test

Just go through the test plan normally and make sure everything continues to work.

### Related issues

- Part of #546

### Backwards compatibility

N/A