merge bundled translations with cached OTA translations on app load (#1565)
### Description

During development you may need to add a new translation key as part of your new feature. While this feature/new translation key is not merged to main, it will not be registered in Crowdin and will therefore not come through as part of the OTA translations. As a result you'll see raw translation keys during development which doesn't give confidence that the feature is working correctly.

In this PR, when loading the resources, use the bundled translations as the base and merge cached OTA translations on top so that we can pick up the OTA changes but also have new translation keys that were added in the app.

Note that this means we always load the bundled resource for the selected language (as opposed to before when we could save this resource load if there are cached translations)

### Other changes

N/A

### Tested

Manually add a new key to the translation.json in feature branch and verify that it is displayed correctly in the app

### How others should test

Same as above

### Related issues

N/A

### Backwards compatibility

N/A