Update Sentry to latest version (#1620)
### Description

This PR updates Sentry to the latest version. According to the migration docs, there were no changes needed from our side (there were very few breaking changes and we are not using the changed features). 

https://docs.sentry.io/platforms/react-native/migration/


### Tested

Rebuilt both iOS and Android apps, manually (created and) triggered an error and verified that it came through in Sentry.


### How others should test

There should be no change to the app behaviour with this PR, so verify that Sentry is working as expected. Errors that you trigger should be received by Sentry as usual.

### Related issues

Prepares for #1461